# docs/swagger.yaml
basePath: /
definitions:
  dto.ConsumerResponse:
    properties:
      birth_date:
        type: string
      birth_place:
        type: string
      created_at:
        type: string
      full_name:
        type: string
      id:
        type: string
      ktp_photo_url:
        type: string
      legal_name:
        type: string
      nik:
        type: string
      salary:
        type: number
      selfie_photo_url:
        type: string
      updated_at:
        type: string
    type: object
  dto.CreateConsumerRequest:
    properties:
      birth_date:
        description: '"YYYY-MM-DD"'
        type: string
      birth_place:
        type: string
      full_name:
        type: string
      ktp_photo_url:
        type: string
      legal_name:
        type: string
      nik:
        type: string
      salary:
        type: number
      selfie_photo_url:
        type: string
    required:
    - full_name
    - legal_name
    - nik
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: API untuk PT-XYZ Multifinance
  title: PT-XYZ Multifinance API
  version: "1.0"
paths:
  /api/v1/consumers:
    post:
      consumes:
      - application/json
      description: Create/register a new consumer
      parameters:
      - description: Consumer Data
        in: body
        name: consumer
        required: true
        schema:
          $ref: '#/definitions/dto.CreateConsumerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ConsumerResponse'
        "400":
          description: invalid request or validation error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Register new consumer
      tags:
      - consumer
  /api/v1/consumers/{id}:
    get:
      description: Get detail of consumer by ID
      parameters:
      - description: Consumer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ConsumerResponse'
        "404":
          description: consumer not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get consumer by ID
      tags:
      - consumer
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"